gmail-integration-service/requirements.txt
fastapi==0.115.12
jinja2
# Kafka client for Python
confluent-kafka==2.10.0

# Redis client
redis==5.3.0

# Gmail API integration
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-oauthlib==1.2.2
oauth2client==4.1.3

# PostgreSQL support via SQLAlchemy
SQLAlchemy==2.0.40
psycopg2-binary==2.9.10

# Logging & utilities
python-dotenv==1.1.0
requests==2.32.3
unzip==1.0.0
uvicorn==0.34.2